crawler/crawler/requirements.txt
redis==3.2.1
pillow==5.4.1
pymongo==3.8.0
PyMySQL==0.9.3
scrapy_redis==0.6.8
scrapy==1.6.0
pyasn1==0.4.1
oauth2client==4.1.3
oauthlib==3.0.0
scrapy_crawlera
google-api-core==1.6.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
google-cloud-core==0.28.1
google-cloud-pubsub==0.40.0
google-cloud-speech==1.1.0
google-cloud-storage==1.13.0
google-cloud-vision==0.35.2
google-resumable-media==0.3.1
googleapis-common-protos==1.5.10



